fix: resolve async component issue with RSS fetching
Properly separate server and client components for error-free async data fetching.

#VERCEL_SKIP
