+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-v0-project",
+  "name": "confer-solutions-ai",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,14 +8,11 @@
     "lint": "next lint",
     "start": "next start",
     "test": "jest",
-    "test:coverage": "jest --coverage",
     "test:watch": "jest --watch"
   },
-  "overrides": {
-    "@testing-library/react": "^16.0.0"
-  },
   "dependencies": {
-    "@hookform/resolvers": "^3.9.1",
+    "@ai-sdk/openai": "^0.0.66",
+    "@hookform/resolvers": "^3.9.0",
     "@radix-ui/react-accordion": "latest",
     "@radix-ui/react-alert-dialog": "latest",
     "@radix-ui/react-aspect-ratio": "latest",
@@ -43,13 +40,17 @@
     "@radix-ui/react-toggle": "latest",
     "@radix-ui/react-toggle-group": "latest",
     "@radix-ui/react-tooltip": "latest",
-    "autoprefixer": "^10.4.20",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "ai": "^3.4.32",
     "canvas-confetti": "latest",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "latest",
-    "date-fns": "4.1.0",
+    "date-fns": "^4.1.0",
     "embla-carousel-react": "latest",
+    "framer-motion": "^11.11.17",
     "input-otp": "latest",
     "lucide-react": "^0.454.0",
     "next": "15.2.4",
@@ -66,17 +67,17 @@
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
     "vaul": "latest",
-    "zod": "^3.24.1"
+    "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.1.5",
-    "@testing-library/react": "^16.0.0",
-    "@types/jest": "^29.5.8",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@types/jest": "^29.5.14",
     "@types/node": "^22",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "eslint": "9.31.0",
-    "eslint-config-next": "15.4.2",
+    "eslint": "^8",
+    "eslint-config-next": "15.1.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.5",
